Created architecture for taxes formulas
