+++ app/models/concerns/nue_formulas.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+module NueFormulas
+  include(IncomeTaxesBaseFormulas)
+
+  def test2
+    'ee2'
+  end
+end
